Agent A: Fix mypy --strict errors in pipeline, metrics, and reporting modules
- Fix datetime.UTC -> datetime.timezone.utc (3 occurrences in __main__.py)
- Add type parameter to _get_available_models return type
- Remove unused type: ignore comments in chunking/__init__.py (6 lines)
- Fix object typing to Any for lazy-loaded adapters in embedding/pipeline.py
- Replace type: ignore with typed variable assignments in pipeline.py
- Wrap pio.to_html() returns in str() in reporting/html.py (5 lines)
- Fix variable shadowing (sm -> sm_entry) in html.py heatmap
- Fix "in" operator on object type in structural.py (isinstance guard)
- Wrap variance**0.5 in float() in structural.py
- Wrap sum() in float() in retrieval.py _dcg()
- Wrap faiss ntotal in int() in retrieval/index.py

Resolves 31 mypy --strict errors in Agent A owned files.
Remaining 1 error is in dashboard/page_retrieval.py (Agent B).

+++ src/scaffolder/metrics/retrieval.py
@@ -123,9 +123,11 @@ def ndcg_at_k(
 
 def _dcg(grades: list[int]) -> float:
     """Discounted Cumulative Gain."""
-    return sum(
-        (2**g - 1) / math.log2(i + 2)  # i+2: 0-indexed i, rank 1-indexed
-        for i, g in enumerate(grades)
+    return float(
+        sum(
+            (2**g - 1) / math.log2(i + 2)  # i+2: 0-indexed i, rank 1-indexed
+            for i, g in enumerate(grades)
+        )
     )
